Avoid caching BroadcastMessage as an ActiveRecord object
When a Rails 4 host serializes a BroadcastMessage, it will serialize
`ActiveRecord::ConnectionAdapters::PostgreSQL::OID::Integer`, which does
not exist in Rails 5. This will cause Error 500s on a Rails 5 reading
from this cache.

To make Rails 4 and 5 play well together, store the data as JSON and
construct the ActiveRecord objects from JSON.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/55034