Don't allow Danger to fail in CI any more
In https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/20654 this job
was marked as allowing failures, but this allows developers to ignore
the results that Danger produces. Since it's not clear why the job was
allowed to fail, and since it's somewhat counter productive, we no
longer allow it to fail.