[QA] Fix the 'Commit data' QA test
We now visit the project/wiki page after a push since that's the logical
page where we should be after a push.

Also, we now properly set the `@project` variable because the
`web_url` of a Push resource isn't necessarily it's project/wiki page
(if the project/wiki are created via the API for instance, `current_url`
would be the current page, not the project/wiki page).

Signed-off-by: Rémy Coutable <remy@rymai.me>