Resolve transient failure in Help page spec
Sometimes due to a slow request to load the version check image, the
placeholder image data was still being seen, resulting in this failure:

    expected "data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw==" to end with "/version-check-url"

Now we check the `data-src` attribute, which allows us to remove the
`:js` metadata.

This commit also removes a redundant test, which was just ensuring that
the selector we were already using in the other test is visible. If this
test were failing, the other one would always fail too, so it was
pointless.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/55372