Add docs on how web terminals are secured
In the administration page for web terminals add a new `Security`
section explaining how the interactive web terminals are secured between
GitLab and gitlab-runner.

This section is under administration instead of
`doc/ci/interactive_web_terminal` because the administrators of the
GitLab instance would be mostly interested in this, whilst
`doc/ci/interactive_web_terminal` are interested in how the feature
works.

closes https://gitlab.com/gitlab-org/gitlab-ce/issues/52681