Don't ignore conflicts in db/schema.rb
It's not entirely clear to me why we were ignoring db/schema.rb in the
first place, but this should not be done now that we merge CE into EE
automatically. Doing so can lead to changes being thrown away in the
event of a conflict, which can then lead to EE pipelines failing.