Fix Bitbucket Server import only including first 25 pull requests
The change to paginate repos in
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/22825 caused the
paginator to stop after 25 pull requests because the limit was set to 25
if none was defined. To fix this, we should only stop if the limit has
actually been set and use the limit parameter to determine the maximum
number of items to process per page.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/55914