Fix QA spec flaky for application secret variables
The QA spec for application secrets fails when run after the auto devops
spec because no pipeline for it is triggered :(

When the QA spec for application has passed, it has always been the
first spec to run hence it is able to set the K8S_SECRET_ variable
before the pipeline triggered by the initial commit has run.

The correct fix would be manually trigger a pipeline after we have added
the K8S_SECRET_* variable. However, because Helm does not update the
Deployment, we need to push a code change to make Helm update the
Deployment and the pods.