Disable audit event logging for pipeline destruction
AuditEventService isn't equipped to handle logging of the destruction of
entities such as CI pipelines. It's a project-level event that operates
on a pipeline. The current log doesn't even indicate that the pipeline
is being destroyed.

This is a CE backport of
https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/9105.  We're
removing the auditing call because it breaks the EE implementation.