[Review Apps] Don't exit when dependable job fails
The function that retrieves the dependable job is pretty basic and
retrieves the first job found with the matching name, but this job can
be failed but then successfully retried. In that case, we would exit the
depending job even though the dependable job actually succeeded (the
second time). Let's simplify things, be optimistic and continue with the
depending job even if the dependable job fails.

That reverts to the original behavior.

Signed-off-by: Rémy Coutable <remy@rymai.me>