Indicate on Status if Issue was Moved
Add the text `(moved)` to the `Closed` status indicator on a closed
issue if the reason the issue was closed was due to moving it.

This only applies to closed issues. Issues that were closed and moved
then later re-opened only show `Open`. This makes it more immidately
clear why the issue was closed.