Fix 404s for snippet uploads when relative URL root used
Personal snippet uploads have neither a group nor a project. If a GitLab
instance were configured with a relative URL root (e.g. `/gitlab`), then
the Markdown filter would not include this root in the generated path.
We fix this by adding this root if there is no group or project.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/56280