Removes duplicated cluster show partials
gcp/show and user/show partial contain basically the same html elements
causing duplication.

The difference between the partials is that one of them has
html elements disabled if the cluster is managed. This MR aims to
remove that duplication by having just a single
show partial for both types