Fix flaky wiki create test
There's an svg on the page that allows you to create a wiki page.
The svg takes a fraction of a second to load after which the
"Create your first page" button shifts up a bit. This can cause
 webdriver to miss the hit so we wait for the svg to load before
clicking the button.

Also update the elements used in the test to conform to our best
practice.

And replace `act` with `perform`

Finally, remove the `before` block and `login` method, making the
code slightly simpler.