Fix pagination and duplicate requests in environments page
This commit fixes two problems:

1. When a user clicked on another page, the app would fetch the new page
data but never update because the request and response parameters failed
to match the isEqual check. This was happening because the JSON response
omitted the `nested` attribute, so there was a comparison with a `null`
value.  We fix this by scrubbing undefined values before doing the
comparison.

2. There were duplicate requests made for the environments page because
the success handler of fetchPipelines() would cause the polling
component to make another XHR request.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/58191