Upgrade jest and related dependencies
This removes other outdated babel@6 dependencies as well.

Unfortunately the newer version of vue-jest has a dependency of an older
version of ts-jest. In order to satisfy the peerDependency from ts-jest,
we are forcing the version@24. The only "breaking" change from ts-jest
23->24 is that it requires a newer version of jest. This might be
obsolete soon, as vue-jest plans on switching to babel eventually.