Freeze date in merge request status view spec
Previously, when this spec ran on the 28th of February, "one month ago"
isn't actually a month ago, it's "28 days ago".

Now we freeze the date to one that works as intended, since the point of
the test is only that it shows a relative date in the tooltip, not
necessarily that it's an exact, specific date.