Show pre-receive error message in merge request widget
When a merge fails due to a locked file, no feedback was given to the
user as to why it failed if the user used the Merge button.  However, if
the push attempt happened directly in the Web UI, the pre-receive
message would be displayed.

To be consistent and to make it clear why a merge failed, we now include
the pre-receive error message in the merge request widget.

Closes https://gitlab.com/gitlab-org/gitlab-ee/issues/10165