Document ChatOps for actor feature gates
The documentations only covered the percentage based feature gates.
These don't work really well for a number of reasons, for example; a new
visual element in GitLab would be weird to see only 25% of the time.

Allowing feature gates based on projects and groups has been supported,
now it's documented.