Fix error creating a merge request when diff includes a null byte
If a diff happened to include a single null byte anywhere, insertion
into the database would fail with an Error 500 since the column is text
and not a byte array. To fix this, we mark the diff as binary if we
detect a single null byte and Base64-encode it.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/57710