Add multi-line suggestion migrations
Adds outdated, lines_above and lines_below columns to
suggestions table.

outdated - boolean which represents whether the suggestion
is outdated or not. For instance, if any line changed after
you left the multi-line suggestion, even though the note
is not outdated, it helps tracking if the content has changed
in the latest file. We cache this information in a column given
it's not a cheap operation to do for every suggestion in the
request time.
lines_below, lines_above - persists the parsed arguments from
`suggestion:-10+3` syntax, where `10` would be lines_above and
3 lines_below. We need that to dynamically calculate which
lines we should monitor for outdating / persisting the correct
content in from_content column.