Use the same standard for method names on project activity and menu
... page objects.

Also, update the affected test case.

I choose `click_...` instead of `go_to_...` since it describes
better what the method really does, which is clicking on something.

I know there are other places in the code where methods start with
`click_` and others where methods start with `go_to_`, with that
I'd like to start a discussion or be pointed to the right issue if
there is already one so that we can address this in the whole
framework.
This contribution could be the start in the direction of using the
standard chosen here.